Renamed coroutine package to threads, added runSafe()
