+++ src/main/java/me/zeroeightsix/kami/gui/mc/KamiGuiDisconnected.kt
@@ -3,7 +3,7 @@ package me.zeroeightsix.kami.gui.mc
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import me.zeroeightsix.kami.module.modules.combat.AutoLog
-import me.zeroeightsix.kami.util.coroutine.mainScope
+import me.zeroeightsix.kami.util.threads.mainScope
 import net.minecraft.client.gui.GuiButton
 import net.minecraft.client.gui.GuiScreen
 import java.time.LocalTime
