+++ src/main/java/me/zeroeightsix/kami/module/modules/player/FastBreak.kt
@@ -7,7 +7,7 @@ import me.zeroeightsix.kami.event.events.SafeTickEvent
 import me.zeroeightsix.kami.mixin.extension.blockHitDelay
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.setting.Settings
-import me.zeroeightsix.kami.util.coroutine.defaultScope
+import me.zeroeightsix.kami.util.threads.defaultScope
 import net.minecraft.network.play.client.CPacketPlayerDigging
 import org.kamiblue.event.listener.listener
